# apps/web/public/journals-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenScience Journal Discovery API",
    "version": "1.0.0",
    "description": "Public journal discovery and immutable editorial interpretations. Original paper DOI identifies the scientific source; OpenScience version URLs identify derivative interpretations. This document covers the public journal extension, not private editorial operations. API requests are not AI citations."
  },
  "servers": [{"url": "/api", "description": "Same-origin web proxy"}],
  "paths": {
...
    "/research/{publicId}": {
      "get": {
        "operationId": "getResearchOverview",
        "parameters": [{"$ref": "#/components/parameters/PublicId"}],
        "responses": {"200": {"description": "Existing research overview. research.latestVersion is a nullable version number. Withdrawn/restricted identities do not grant access to content.", "content": {"application/json": {"schema": {"type": "object", "properties": {"research": {"type": "object", "additionalProperties": true}}}}}}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/research/{publicId}/v/{versionNo}": {
      "get": {
        "operationId": "getFixedResearchVersion",
        "parameters": [{"$ref": "#/components/parameters/PublicId"}, {"name": "versionNo", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 1}}],
        "responses": {"200": {"description": "Existing fixed-version response. research.journalPackage is optional and present only for a publicly accessible journal interpretation; unavailable content must not be treated as complete research.", "content": {"application/json": {"schema": {"type": "object", "properties": {"research": {"type": "object", "additionalProperties": true, "properties": {"journalPackage": {"oneOf": [{"$ref": "#/components/schemas/JournalPackage"}, {"type": "null"}]}}}}}}}}, "404": {"$ref": "#/components/responses/NotFound"}, "429": {"$ref": "#/components/responses/RateLimited"}}
      }
    }
  },
  "components": {
    "parameters": {
...
      "PublicId": {"name": "publicId", "in": "path", "required": true, "schema": {"type": "string"}}
    },
    "responses": {
      "NotFound": {"description": "Unknown or unpublished identity. Private records are not disclosed."},
      "RateLimited": {"description": "Platform rate limit reached. Respect the response Retry-After header when provided."}
    },
    "schemas": {
...
      "Metadata": {"type": "object", "properties": {"title": {"type": "string"}, "doi": {"type": "string"}, "authors": {"type": "array", "items": {"type": "string"}}, "publishedDate": {"type": "string"}, "journalTitle": {"type": "string"}, "issns": {"type": "array", "items": {"type": "string"}}, "originalUrl": {"type": "string", "format": "uri"}}},
...
      "Evidence": {"type": "object", "required": ["quote", "locator"], "properties": {"quote": {"type": "string"}, "locator": {"type": "string"}}},
      "JournalPackage": {"type": "object", "required": ["metadata", "draft", "source", "license", "review", "versionNo", "url"], "properties": {
        "metadata": {"$ref": "#/components/schemas/Metadata"},
        "identity": {"const": "journal-authored-interpretation-of-published-work"},
        "license": {"type": "string"}, "versionNo": {"type": "integer"}, "url": {"type": "string"},
        "source": {"type": "object", "properties": {"kind": {"enum": ["abstract", "fulltext"]}, "url": {"type": "string", "format": "uri"}, "label": {"type": "string"}, "text": {"type": "string", "description": "Optional. Requires separate permission to publish the source."}}},
        "review": {"type": "object", "properties": {"method": {"const": "editorial-human-review"}, "revision": {"type": "integer"}}},
        "draft": {"type": "object", "properties": {
          "summary": {"type": "string"}, "scope": {"enum": ["abstract", "fulltext"]}, "language": {"enum": ["zh", "en"]},
          "core": {"type": "object", "required": ["problem", "insight", "method", "results", "limitations", "reproducibility"], "additionalProperties": {"type": "string"}},
          "claims": {"type": "array", "items": {"type": "object", "properties": {"text": {"type": "string"}, "kind": {"enum": ["experimental", "simulation", "theoretical", "review", "other"]}, "evidence": {"$ref": "#/components/schemas/Evidence"}}}},
          "figures": {"type": "array", "items": {"type": "object", "properties": {"label": {"type": "string"}, "purpose": {"type": "string"}, "finding": {"type": "string"}, "evidence": {"$ref": "#/components/schemas/Evidence"}}}},
          "faq": {"type": "array", "items": {"type": "object", "properties": {"question": {"type": "string"}, "answer": {"type": "string"}, "evidence": {"$ref": "#/components/schemas/Evidence"}}}}
        }}
      }}
    }
  }
}